refactor: streamline Tailwind CSS integration for v4
- Removed deprecated functions related to Tailwind CSS v3.
- Introduced new utility functions for adding variables, merging themes, and handling custom variants.
- Enhanced the setup process to check for missing custom variants and log them appropriately.
- Improved variable management within the @layer blocks for better compatibility with Tailwind CSS v4.
